check message waiting issue

+++ app/Console/Commands/RunMessageQueue.php
@@ -26,6 +26,12 @@ class RunMessageQueue extends Command
      * @var string
      */
     protected $description = 'Command description';
+    
+    // custom defined vars
+    const WAITING_MESSAGE_LIMIT = 100;
+    
+    // waiting messages group
+    public $waitingMessages = [];
 
     /**
      * Create a new command instance.
@@ -55,32 +61,65 @@ class RunMessageQueue extends Command
 
         if ($time->between($morning, $evening, true)) {
             // Get groups
-            $groups = DB::table('message_queues')->distinct()->get(['group_id']);
+            $groups = DB::table('message_queues')->groupBy("group_id")->select("group_id")->get(['group_id']);
 
             foreach ($groups as $group) {
                 // Get messages
-                $message_queues = MessageQueue::where('group_id', $group->group_id)->where('sending_time', '<=', Carbon::now())->where('sent', 0)->where('status', '!=', 1)->orderBy('sending_time', 'ASC')->limit(20);
+                $message_queues = MessageQueue::where('group_id', $group->group_id)
+                ->where('sending_time', '<=', Carbon::now())
+                ->where('sent', 0)
+                ->where('status', '!=', 1)
+                ->orderBy('sending_time', 'ASC')
+                ->limit(20);
+            
+                $findUniqueNumber =  clone $message_queues;
+                $findUniqueNumber = $findUniqueNumber->groupby("whatsapp_number")->select("whatsapp_number")->get();
+
+                $this->waitingMessages = [];
+                if(!$findUniqueNumber->isEmpty()) {
+                    foreach($findUniqueNumber as $uniqueNumber) {
+                        $number                         = !empty($uniqueNumber->whatsapp_number) ? (string)$uniqueNumber->whatsapp_number : 0;
+                        $waitingMessage                 = $this->waitingLimit($number);
+                        $this->waitingMessages[$number] = $waitingMessage;
+                    }
+                }
 
                 // Do we have results?
                 if (count($message_queues->get()) > 0) {
                     foreach ($message_queues->get() as $message) {
-                        if ($message->type == 'message_all') {
 
-                            $customer = Customer::find($message->customer_id);
+                        // check message can able to send 
+                        $number = !empty($message->whatsapp_number) ? (string)$message->whatsapp_number : 0;
+                        
+                        if(!$this->isWaitingFull($number)) {
+                            if ($message->type == 'message_all') {
 
-                            if ($customer && $customer->do_not_disturb == 0) {
-                                SendMessageToAll::dispatchNow($message->user_id, $customer, json_decode($message->data, true), $message->id);
+                                $customer = Customer::find($message->customer_id);
 
-                                dump('sent to all');
+                                if ($customer && $customer->do_not_disturb == 0) {
+                                    SendMessageToAll::dispatchNow($message->user_id, $customer, json_decode($message->data, true), $message->id);
+
+                                    dump('sent to all');
+                                } else {
+                                    $message->delete();
+
+                                    dump('deleting queue');
+                                }
                             } else {
-                                $message->delete();
+                                SendMessageToSelected::dispatchNow($message->phone, json_decode($message->data, true), $message->id, $message->whatsapp_number);
 
-                                dump('deleting queue');
+                                dump('sent to selected');
                             }
-                        } else {
-                            SendMessageToSelected::dispatchNow($message->phone, json_decode($message->data, true), $message->id, $message->whatsapp_number);
 
-                            dump('sent to selected');
+                            // start to add more if there is existing already
+                            if(isset($this->waitingMessages[$number])) {
+                                $this->waitingMessages[$number] = $this->waitingMessages[$number] + 1;
+                            }else{
+                                $this->waitingMessages[$number] = 1;
+                            }
+
+                        }else{
+                            dump('Due to more waiting message for given number : '.$number );
                         }
                     }
                 }
@@ -90,5 +129,85 @@ class RunMessageQueue extends Command
         }
 
         $report->update(['end_time' => Carbon:: now()]);
+    }
+
+    /**
+     * Check waiting is full for given number
+     *
+     */
+
+    private function isWaitingFull($number)
+    {
+        $number = !empty($number) ? $number : 0;
+
+        if(isset($this->waitingMessages[$number]) && $this->waitingMessages[$number] > self::WAITING_MESSAGE_LIMIT) {
+            return true;
+        }
+
+        return false;
+
+    }
+
+    /**
+     * Get instance from whatsapp number
+     * 
+     */
+
+    private function getInstance($number = null)
+    {
+        $number = !empty($number) ? $number : 0;
+
+        return isset(config("apiwha.instances")[$number]) 
+        ? config("apiwha.instances")[$number] 
+        : config("apiwha.instances")[0];
+       
+    }
+
+    /**
+     * send request for find waiting message number
+     * 
+     */
+
+    private function waitingLimit($number = null)
+    {
+        $instance   = $this->getInstance($number);
+        $instanceId = isset($instance["instance_id"]) ? $instance["instance_id"] : 0;
+        $token      = isset($instance["token"]) ? $instance["token"] : 0;
+
+        $waiting = 0;
+
+        if(!empty($instanceId) && !empty($token)) {
+            // executing curl
+            $curl = curl_init();
+        
+            curl_setopt_array($curl, array(
+                CURLOPT_URL => "https://api.chat-api.com/instance$instanceId/showMessagesQueue?token=$token",
+                CURLOPT_RETURNTRANSFER => true,
+                CURLOPT_ENCODING => "",
+                CURLOPT_MAXREDIRS => 10,
+                CURLOPT_TIMEOUT => 300,
+                CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,
+                CURLOPT_HTTPHEADER => array(
+                    "content-type: application/json"
+                ),
+            ));
+
+            $response = curl_exec($curl);
+            $err = curl_error($curl);
+            curl_close($curl);
+
+            if ($err) {
+                // throw some error if you want
+            } else {
+                $result = json_decode($response, true);
+                if(isset($result["totalMessages"]) && is_numeric($result["totalMessages"])){
+                    $waiting = $result["totalMessages"];
+                }
+            }
+            
+        }
+
+        return $waiting;
+        
     }
 }
